requirements.txt
datasets>=2.14.0
huggingface-hub>=0.16.0
openai>=1.0.0
qdrant-client>=1.6.0
networkx>=3.1
py2neo>=2021.2.3
langchain>=0.1.0
langgraph>=0.0.40
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
anyio>=3.7.1
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0
streamlit>=1.28.0
python-dotenv>=1.0.0
